Рефакторит класс TSVFileReader
