+++ src/common/file-reader/file-reader.interface.ts
@@ -1,7 +1,4 @@
-import { Offer } from '../../types/offer.type';
-
 export interface FileReaderInterface {
 	readonly filename: string;
 	read(): void;
-  toArray(): Offer[];
 }
