+++ src/common/file-reader/tsv-file-reader.ts
@@ -1,59 +1,36 @@
-import { readFileSync } from 'fs';
-import { Offer } from '../../types/offer.type.js';
+import EventEmitter from 'events';
+import { createReadStream } from 'fs';
 import { FileReaderInterface } from './file-reader.interface.js';
 
-export default class TSVFileReader implements FileReaderInterface {
-  private rawData = '';
-
-  constructor(public filename: string) {}
-
-  public read(): void {
-    this.rawData = readFileSync(this.filename, { encoding: 'utf8' });
+export default class TSVFileReader extends EventEmitter implements FileReaderInterface {
+  constructor(public filename: string) {
+    super();
   }
 
-  public toArray(): Offer[] {
-    if (!this.rawData) {
-      return [];
+  public async read():Promise<void> {
+    const stream = createReadStream(this.filename, {
+      highWaterMark: 16384,
+      encoding: 'utf-8',
+    });
+
+    console.log('Stream - ', stream);
+
+    let lineRead = '';
+    let endLinePosition = -1;
+    let importedRowCount = 0;
+
+    for await (const chunk of stream) {
+      lineRead += chunk.toString();
+
+      while ((endLinePosition = lineRead.indexOf('\n')) >= 0) {
+        const completeRow = lineRead.slice(0, endLinePosition + 1);
+        lineRead = lineRead.slice(++endLinePosition);
+        importedRowCount++;
+
+        this.emit('line', completeRow);
+      }
     }
 
-    return this.rawData
-      .split('\n')
-      .filter((row) => row.trim() !== '')
-      .map((line) => line.split('\t'))
-      .map(([bedrooms, cityName, cityLatitude, cityLongitude, cityZoom, description, goods, hostAvatar, hostId, hostStatus, hostName, offerId, images, isFavorite, isPremium, offerLatitude, offerLongitude, offerZoom, maxAdults, postDate, previewImage, price, rating, title, type]) => ({
-        bedrooms: Number(bedrooms),
-        city: {
-          name: cityName,
-          location: {
-            latitude: Number(cityLatitude),
-            longitude: Number(cityLongitude),
-            zoom: Number(cityZoom)
-          }
-        },
-        description: description,
-        goods: goods.split(';'),
-        host: {
-          avatarUrl: hostAvatar,
-          id: Number(hostId),
-          isPro: Boolean(Number(hostStatus)),
-          name: hostName
-        },
-        id: Number(offerId),
-        images: images.split(';'),
-        isFavorite: Boolean(Number(isFavorite)),
-        isPremium: Boolean(Number(isPremium)),
-        location: {
-          latitude: Number(offerLatitude),
-          longitude: Number(offerLongitude),
-          zoom: Number(offerZoom)
-        },
-        maxAdults: Number(maxAdults),
-        postDate: new Date(postDate),
-        previewImage,
-        price: Number(price),
-        rating: Number(rating),
-        title,
-        type
-      }));
+    this.emit('end', importedRowCount);
   }
 }
