+++ src/utils/common.ts
@@ -1,0 +1,42 @@
+import { Offer } from '../types/offer.type.js';
+
+export const createOffer = (row: string) => {
+  const tokens = row.replace('\n', '').split('\t');
+  const [bedrooms, cityName, cityLatitude, cityLongitude, cityZoom, description, goods, hostAvatar, hostId, hostStatus, hostName, offerId, images, isFavorite, isPremium, offerLatitude, offerLongitude, offerZoom, maxAdults, postDate, previewImage, price, rating, title, type] = tokens;
+
+  return {
+    bedrooms: Number(bedrooms),
+    city: {
+      name: cityName,
+      location: {
+        latitude: Number(cityLatitude),
+        longitude: Number(cityLongitude),
+        zoom: Number(cityZoom)
+      }
+    },
+    description: description,
+    goods: goods.split(';'),
+    host: {
+      avatarUrl: hostAvatar,
+      id: Number(hostId),
+      isPro: Boolean(Number(hostStatus)),
+      name: hostName
+    },
+    id: Number(offerId),
+    images: images.split(';'),
+    isFavorite: Boolean(Number(isFavorite)),
+    isPremium: Boolean(Number(isPremium)),
+    location: {
+      latitude: Number(offerLatitude),
+      longitude: Number(offerLongitude),
+      zoom: Number(offerZoom)
+    },
+    maxAdults: Number(maxAdults),
+    postDate: new Date(postDate),
+    previewImage,
+    price: Number(price),
+    rating: Number(rating),
+    title,
+    type
+  } as Offer;
+};
